+++ flow_runner.go
@@ -3,7 +3,6 @@ package taskflow
 import (
 	"context"
 	"errors"
-	"flag"
 	"fmt"
 	"io"
 	"sort"
@@ -14,79 +13,79 @@ import (
 
 type flowRunner struct {
 	output      io.Writer
-	params      Params
+	params      map[string]paramValueFactory
+	paramValues map[string]ParamValue
 	tasks       map[string]Task
-	verbose     bool
+	verbose     BoolParam
 	defaultTask RegisteredTask
 }
 
 // Run runs provided tasks and all their dependencies.
 // Each task is executed at most once.
 func (f *flowRunner) Run(ctx context.Context, args []string) int {
-	// prepare flag.FlagSet
-	cli := flag.NewFlagSet("", flag.ContinueOnError)
-	cli.SetOutput(f.output)
-	verbose := cli.Bool("v", false, "Verbose output: log all tasks as they are run. Also print all text from Log and Logf calls even if the task succeeds.")
-	usage := func() {
-		fmt.Fprintf(cli.Output(), "Usage: [flag(s)] [key=val] task(s)\n")
-		fmt.Fprintf(cli.Output(), "Flags:\n")
-		cli.PrintDefaults()
-
-		fmt.Fprintf(cli.Output(), "Tasks:\n")
-		w := tabwriter.NewWriter(cli.Output(), 1, 1, 4, ' ', 0)
-		keys := make([]string, 0, len(f.tasks))
-		for k, task := range f.tasks {
-			if task.Description == "" {
-				continue
-			}
-			keys = append(keys, k)
-		}
-		sort.Strings(keys)
-		for _, k := range keys {
-			t := f.tasks[k]
-			fmt.Fprintf(w, "  %s\t%s\n", t.Name, t.Description)
-		}
-		w.Flush() //nolint // not checking errors when writting to output
-
-		if f.defaultTask.name != "" {
-			fmt.Fprintf(cli.Output(), "Default task: %s\n", f.defaultTask.name)
-		}
-
-		if len(f.params) > 0 {
-			fmt.Fprintf(cli.Output(), "Default parameters:\n")
-			for key, val := range f.params {
-				fmt.Fprintf(w, "  %s\t%s\n", key, val)
-			}
-			w.Flush() //nolint // not checking errors when writting to output
-		}
-	}
-	cli.Usage = usage
-
-	// parse args (flags)
-	if err := cli.Parse(args); err != nil {
-		return CodeInvalidArgs
-	}
-	if *verbose {
-		f.verbose = true
+	if unusedParams := f.unusedParams(); len(unusedParams) > 0 {
+		panic(fmt.Sprintf("unused parameters: %v\n", unusedParams))
 	}
 
-	// parse non-flag args (tasks and parameters)
+	f.paramValues = make(map[string]ParamValue)
+	for _, param := range f.params {
+		value := param.newValue()
+		f.paramValues[param.info.Name] = value
+	}
+	usageRequested := false
+
+	var argHandler func(string) error
+
+	handleNextArgFor := func(value ParamValue) {
+		nextHandler := argHandler
+		argHandler = func(s string) error {
+			err := value.Set(s)
+			argHandler = nextHandler
+			return err
+		}
+	}
+
 	var tasks []string
-	for _, arg := range cli.Args() {
-		if paramAssignmentIdx := strings.IndexRune(arg, '='); paramAssignmentIdx > 0 {
-			// parameter assignement via 'key=val'
-			key := arg[0:paramAssignmentIdx]
-			val := arg[paramAssignmentIdx+1:]
-			f.params[key] = val
-			continue
+
+	argHandler = func(arg string) error {
+		if _, isTask := f.tasks[arg]; isTask {
+			tasks = append(tasks, arg)
+			return nil
 		}
-		if _, ok := f.tasks[arg]; !ok {
-			// task is not registered
-			fmt.Fprintf(f.output, "task provided but not registered: %s\n", arg)
-			usage()
+		if arg[0] == '-' {
+			// parase parameters
+			split := strings.SplitN(arg[1:], "=", 2)
+			if value, isFlag := f.paramValues[split[0]]; isFlag {
+				switch {
+				case len(split) > 1:
+					return value.Set(split[1])
+				case value.IsBool():
+					return value.Set("")
+				default:
+					handleNextArgFor(value)
+					return nil
+				}
+			}
+		}
+		// if they haven't been overridden above, provide usage for common queries
+		if (arg == "-h") || (arg == "--help") || (arg == "help") {
+			usageRequested = true
+			return nil
+		}
+		fmt.Fprintf(f.output, "unknown argument: %s\n", arg)
+		return fmt.Errorf("unknown argument: %s", arg)
+	}
+
+	for _, arg := range args {
+		err := argHandler(arg)
+		if err != nil {
 			return CodeInvalidArgs
 		}
-		tasks = append(tasks, arg)
+	}
+
+	if usageRequested {
+		printUsage(f)
+		return CodePass
 	}
 
 	// set default task if none is provided
@@ -95,21 +94,24 @@ func (f *flowRunner) Run(ctx context.Context, args []string) int {
 	}
 
 	if len(tasks) == 0 {
-		fmt.Fprintln(cli.Output(), "no task provided")
-		usage()
+		fmt.Fprintln(f.output, "no task provided")
+		printUsage(f)
 		return CodeInvalidArgs
 	}
 
-	// recursive run
+	return f.runTasks(ctx, tasks)
+}
+
+func (f *flowRunner) runTasks(ctx context.Context, tasks []string) int {
 	from := time.Now()
 	executedTasks := map[string]bool{}
 	for _, name := range tasks {
 		if err := f.run(ctx, name, executedTasks); err != nil {
-			fmt.Fprintf(cli.Output(), "%v\t%.3fs\n", err, time.Since(from).Seconds())
+			fmt.Fprintf(f.output, "%v\t%.3fs\n", err, time.Since(from).Seconds())
 			return CodeFailure
 		}
 	}
-	fmt.Fprintf(cli.Output(), "ok\t%.3fs\n", time.Since(from).Seconds())
+	fmt.Fprintf(f.output, "ok\t%.3fs\n", time.Since(from).Seconds())
 	return CodePass
 }
 
@@ -118,7 +120,7 @@ func (f *flowRunner) run(ctx context.Context, name string, executed map[string]b
 	if executed[name] {
 		return nil
 	}
-	for _, dep := range task.Dependencies {
+	for _, dep := range task.Deps {
 		if err := f.run(ctx, dep.name, executed); err != nil {
 			return err
 		}
@@ -142,37 +144,123 @@ func (f *flowRunner) runTask(ctx context.Context, task Task) bool {
 		return true
 	}
 
-	w := f.output
-	if !f.verbose {
-		w = &strings.Builder{}
+	paramValues := make(map[string]ParamValue)
+	for _, param := range task.Params {
+		paramValues[param.Name()] = f.paramValues[param.Name()]
 	}
 
-	// report task start
-	fmt.Fprintf(w, "===== TASK  %s\n", task.Name)
+	// if verbose flag is registered then check its value
+	verboseParamVal, ok := f.paramValues[f.verbose.Name()]
+	verbose := ok && verboseParamVal.Get().(bool)
 
-	// run task
-	runner := Runner{
-		Ctx:      ctx,
-		TaskName: task.Name,
-		Verbose:  f.verbose,
-		Params:   f.params,
-		Output:   w,
+	failed := false
+	measuredCommand := func(tf *TF) {
+		w := tf.Output()
+		if !verbose {
+			w = &strings.Builder{}
+		}
+
+		// report task start
+		fmt.Fprintf(w, "===== TASK  %s\n", tf.Name())
+
+		// run task
+		runner := Runner{
+			Ctx:         tf.Context(),
+			TaskName:    tf.Name(),
+			ParamValues: tf.paramValues,
+			Output:      w,
+		}
+		result := runner.Run(task.Command)
+
+		// report task end
+		status := "PASS"
+		switch {
+		case result.Failed():
+			status = "FAIL"
+			failed = true
+		case result.Skipped():
+			status = "SKIP"
+		}
+		fmt.Fprintf(w, "----- %s: %s (%.2fs)\n", status, tf.Name(), result.Duration().Seconds())
+
+		if sb, ok := w.(*strings.Builder); ok && result.failed {
+			io.Copy(tf.Output(), strings.NewReader(sb.String())) //nolint // not checking errors when writting to output
+		}
 	}
-	result := runner.Run(task.Command)
 
-	// report task end
-	status := "PASS"
-	switch {
-	case result.Failed():
-		status = "FAIL"
-	case result.Skipped():
-		status = "SKIP"
+	measuredRunner := Runner{
+		Ctx:         ctx,
+		TaskName:    task.Name,
+		ParamValues: paramValues,
+		Output:      f.output,
 	}
-	fmt.Fprintf(w, "----- %s: %s (%.2fs)\n", status, task.Name, result.Duration().Seconds())
+	measuredRunner.Run(measuredCommand)
 
-	if sb, ok := w.(*strings.Builder); ok && result.failed {
-		io.Copy(f.output, strings.NewReader(sb.String())) //nolint // not checking errors when writting to output
+	return !failed
+}
+
+func (f *flowRunner) unusedParams() []string {
+	remainingParams := make(map[string]struct{})
+	for key := range f.params {
+		remainingParams[key] = struct{}{}
+	}
+	delete(remainingParams, f.verbose.Name())
+	for _, task := range f.tasks {
+		for _, param := range task.Params {
+			delete(remainingParams, param.Name())
+		}
+	}
+	unusedParams := make([]string, 0, len(remainingParams))
+	for key := range remainingParams {
+		unusedParams = append(unusedParams, key)
+	}
+	return unusedParams
+}
+
+func printUsage(f *flowRunner) {
+	flagName := func(paramName string) string {
+		return "-" + paramName
 	}
 
-	return !result.failed
+	fmt.Fprintf(f.output, "Usage: [flag(s) | task(s)]...\n")
+	fmt.Fprintf(f.output, "Flags:\n")
+	w := tabwriter.NewWriter(f.output, 1, 1, 4, ' ', 0)
+	keys := make([]string, 0, len(f.params))
+	for key := range f.params {
+		keys = append(keys, key)
+	}
+	sort.Strings(keys)
+	for _, key := range keys {
+		param := f.params[key]
+		fmt.Fprintf(w, "  %s\tDefault: %s\t%s\n", flagName(param.info.Name), param.newValue().String(), param.info.Usage)
+	}
+	w.Flush() //nolint // not checking errors when writing to output
+
+	fmt.Fprintf(f.output, "Tasks:\n")
+	keys = make([]string, 0, len(f.tasks))
+	for k, task := range f.tasks {
+		if task.Usage == "" {
+			continue
+		}
+		keys = append(keys, k)
+	}
+	sort.Strings(keys)
+	for _, k := range keys {
+		t := f.tasks[k]
+		params := make([]string, len(t.Params))
+		for i, param := range t.Params {
+			params[i] = flagName(param.Name())
+		}
+		sort.Strings(params)
+		paramsText := ""
+		if len(params) > 0 {
+			paramsText = "; " + strings.Join(params, " ")
+		}
+		fmt.Fprintf(w, "  %s\t%s%s\n", t.Name, t.Usage, paramsText)
+	}
+	w.Flush() //nolint // not checking errors when writing to output
+
+	if f.defaultTask.name != "" {
+		fmt.Fprintf(f.output, "Default task: %s\n", f.defaultTask.name)
+	}
 }